# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para plataforma de serviços de limpeza.",
        "title": "Phresh-Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/limpezas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limpezas"
                ],
                "summary": "Listar catálogo de limpezas",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número da página",
                        "name": "pagina",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Itens por página",
                        "name": "tamanho",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaLimpeza"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limpezas"
                ],
                "summary": "Criar novo serviço de limpeza",
                "parameters": [
                    {
                        "description": "Dados da limpeza",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoCriarLimpeza"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaLimpeza"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/limpezas/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limpezas"
                ],
                "summary": "Buscar limpeza por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaLimpeza"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limpezas"
                ],
                "summary": "Atualizar serviço de limpeza",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da limpeza",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoAtualizarLimpeza"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaLimpeza"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limpezas"
                ],
                "summary": "Deletar serviço de limpeza",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
...
        "/usuarios/eu/limpezas": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limpezas"
                ],
                "summary": "Listar limpezas do faxineiro autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaLimpeza"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.RequisicaoAtualizarLimpeza": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "duracao_estimada": {
                    "type": "number"
                },
                "nome": {
                    "type": "string"
                },
                "tipo_limpeza": {
                    "type": "string"
                },
                "valor_hora": {
                    "type": "number"
                }
            }
        },
...
        "dto.RequisicaoCriarLimpeza": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "duracao_estimada": {
                    "type": "number"
                },
                "nome": {
                    "type": "string"
                },
                "tipo_limpeza": {
                    "type": "string"
                },
                "valor_hora": {
                    "type": "number"
                }
            }
        },
...
        "dto.RespostaErro": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "detalhes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mensagem": {
                    "type": "string"
                }
            }
        },
...
        "dto.RespostaLimpeza": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "duracao_estimada": {
                    "type": "number"
                },
                "faxineiro_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "preco_total": {
                    "type": "number"
                },
                "tipo_limpeza": {
                    "type": "string"
                },
                "valor_hora": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}